# [WIP] Add a lookup cache for txOutputs