TxOutputs are requested very often and the previous iterations
accumulate lot of cpu time as profiler shows.