@chimp1984 2020-12-15
Here are performance test results.

PR version:
100 calls of daoStateService.getTxOutputsByTxOutputType(TxOutputType.BSQ_OUTPUT) took 2508 ms 
100000 calls of daoStateService.existsTxOutput(TxOutputKey) took 2 ms 
100000 calls of daoStateService.getTxOutput(TxOutputKey) took 3 ms 

Master:
100 calls of daoStateService.getTxOutputsByTxOutputType(TxOutputType.BSQ_OUTPUT) took 4206 ms 
100 calls of daoStateService.existsTxOutput(TxOutputKey) took 2907 ms 
100 calls of daoStateService.existsTxOutput(TxOutputKey) took 2946 ms 

The daoStateService.getTxOutputsByTxOutputType call takes about 30 ms which is very expensive and 
requires more improvement but as it iterates the cache map it is expected to has little impact.

The other 2 methods havea huge difference.
Note that iterations have been increased from 100 to 100000 to get some meaningful time log.

I think we should measure all methods and find more hotspots...

As its a risky domain, I leave it to maintainers to merge the PR or delay it for later.