@chimp1984 2020-12-15
I checked how often the methods are called:
getTxOutputsByTxOutputType is called about 500 times after a few minutes and clicking 
though most DAO screens resulting in 15 sec accululated CPU time.

getTxOutput and existsTxOutput is only used form the BsqCoinSelector in edge cases, 
so only relevant for users who have BSQ in their wallet. I got existsTxOutput called 
3 times and getTxOutput called 0 times with a live app with BSQ.

So unfortunately the big improvements will not have much impact as not called often 
and the one which is called more often has not so much change. 
So not sure if the added risk and the added memory footprint by adding a cache as 
additional data structure (150k objects) are worth it. I would tend to postpone that 
improvement to find the more severe bottlenecks. (I am still a bit surprised as profiler 
reported the getTxOutputsByTxOutputType method as bottleneck)